Bugfix + optimize [semver:minor] (#48)
Address a few known bugs and improve/optimize the config logic, and install logic.

Minor changes:

reconstructed skip-install-check parameter
New bash optimization required re-doing the parameter. Should work essentially the same but with this much change to the logic I did not want to consider this a patch in the case this might affect a user differently. Settling on a minor change, as this shouldn't cause errors.
Bug Fix:

Previously, if you had selected to install AWS CLI v2, but v1 was already installed, the orb would exit without error and skip the install thinking it was already installed.
Orb now checks to see if the installed version matched the config version, if they do not match, the installed version will be replaced.
#47 Fix setting default AWS region
#42 updated tests show success for both versions of CLI on each executor (windows not included)
Other:

Rework install logic
Now checks to see if AWS is already installed, and what version.
Skips if the version installed matches the desired version. If not, the installed version will be replaced automatically.
#45 Provide better error message when attempting to install v1 on a non-python image
Increased and improved testing.