Allow override without specifying version
In the `install` command, if `override_installed` is set to `true` but `version` is not specified, the orb will attempt to download the AWS CLI from a path including "latest". This won't work; you access the latest version by not specifying a version at all in the URL, ie an empty string.