Update image versions
This required updating the downloaded version of aws-cli to 2.13 for
Alpine. You can still specify other versions for other OSes, but
earlier versions of aws-cli would consistently segfault, and so we've
elected to support only the version bundled with the version of Alpine
that is specified, through this orb. Something to keep in mind if you
have specific needs.